Ufp 517 update groups list api

+++ app/src/main/java/io/unityfoundation/dds/permissions/manager/model/group/GroupService.java
@@ -4,6 +4,7 @@ import io.micronaut.data.model.Page;
 import io.micronaut.data.model.Pageable;
 import io.micronaut.http.annotation.Body;
 import io.micronaut.security.authentication.Authentication;
+import io.micronaut.security.authentication.AuthenticationException;
 import io.micronaut.security.utils.SecurityService;
 import io.unityfoundation.dds.permissions.manager.model.user.Role;
 import io.unityfoundation.dds.permissions.manager.model.user.User;
@@ -32,10 +33,23 @@ public class GroupService {
     }
 
     public Page<Group> findAll(Pageable pageable) {
-        return groupRepository.findAll(pageable);
+        Authentication authentication = securityService.getAuthentication().get();
+
+        boolean isAdmin = authentication.getRoles().contains(Role.ADMIN.toString());
+        if (isAdmin) {
+            return groupRepository.findAll(pageable);
+        } else {
+            String userEmail = authentication.getName();
+            User user = userService.getUserByEmail(userEmail).get();
+            return groupRepository.findIfMemberOfGroup(user.getId(), pageable);
+        }
     }
 
     public void save(Group group) {
+        if (!isCurrentUserAdmin()) {
+            throw new AuthenticationException("Not authorized");
+        }
+
         if (group.getId() == null) {
             groupRepository.save(group);
         } else {
@@ -44,6 +58,10 @@ public class GroupService {
     }
 
     public void deleteById(Long id) {
+        if (!isCurrentUserAdmin()) {
+            throw new AuthenticationException("Not authorized");
+        }
+
         groupRepository.deleteById(id);
     }
 
@@ -102,9 +120,13 @@ public class GroupService {
         Authentication authentication = securityService.getAuthentication().get();
         String userEmail = authentication.getName();
 
-        boolean isAdmin = authentication.getRoles().contains(Role.ADMIN.toString());
         boolean isGroupAdmin = group.get().getAdmins().stream().anyMatch(groupAdmins -> groupAdmins.getEmail().equals(userEmail));
 
-        return isAdmin || isGroupAdmin;
+        return isCurrentUserAdmin() || isGroupAdmin;
+    }
+
+    public boolean isCurrentUserAdmin() {
+        Authentication authentication = securityService.getAuthentication().get();
+        return authentication.getRoles().contains(Role.ADMIN.toString());
     }
 }
